update file json for vercel
Another Vercel account is using this domain. Set the following TXT record on _vercel.ztl.is-a-good.dev to use ztl.is-a-good.dev in this project. If you wish to transfer the entire domain, this will need to be done by the current domain owner. Once the verification is completed and the domain is successfully configured, the TXT record can be removed.